Fix/send token with request

+++ src/web-service/frontend/src/lib/shopping-list/ShoppingListEntry.svelte
@@ -3,6 +3,7 @@
     import Checkbox from "$lib/forms/Checkbox.svelte";
     import {handleErrors} from "../../assets/helper/handleErrors";
     import Trash from "../../assets/svg/Trash.svelte";
+    import {fetchHelper} from "../../assets/helper/fetchHelper";
 
     type ViewState = "detailed" | "compressed";
 
@@ -11,27 +12,24 @@
     export let product: { id: number, description: string, ean: number };
     export let view: ViewState = "detailed";
 
-    let price: { price: number } = { price: 0 };
+    let price: any;
     let merchant: string = 'Aldi';
-
-    const apiUrlPrice = `/api/v1/price/${product.id}/2`;
+    const token: string | null = sessionStorage.getItem('access_token');
 
     const dispatch = createEventDispatcher();
 
     onMount(async () => {
-        fetch(apiUrlPrice)
-            .then(handleErrors)
-            .then(data => price = data)
-            .catch(error => console.error("Failed to fetch data:", error.message));
+        const apiUrl: string = `/api/v1/price/${product.id}/${sessionStorage.getItem('user_id')}`;
+        price = await fetchHelper(apiUrl);
     });
 
     function updateShoppingListEntry(): void {
-        if (! listId || ! product.id ) return;
+        if (! listId || ! product.id || ! token) return;
 
         const apiUrl: string = `/api/v1/shoppinglistentries/${listId}/${product.id}`;
         const requestOptions = {
             method: "PUT",
-            headers: { 'Content-Type': 'application/json' },
+            headers: { 'Authorization': `Bearer ${token}` },
             body: `{ "count": ${entry.count}, "checked": ${entry.checked} }`,
         };
 
@@ -42,12 +40,12 @@
     }
 
     function deleteShoppingListEntry(): void {
-        if (! listId || ! product.id) return;
+        if (! listId || ! product.id || ! token) return;
 
         const apiUrl: string = `/api/v1/shoppinglistentries/${listId}/${product.id}`;
         const requestOptions = {
             method: "DELETE",
-            headers: { 'Content-Type': 'application/json' },
+            headers: { 'Authorization': `Bearer ${token}` },
         };
 
         fetch(apiUrl, requestOptions)
@@ -74,11 +72,11 @@
             <Trash classes="w-4 h-4 md:w-5 md:h-5" />
         </button>
     </div>
-    {#if view === 'detailed' && price.price}
+    {#if view === 'detailed' && price}
         <p class="text-gray-dark mt-1 ml-[2.1rem] text-sm flex flex-wrap items-center gap-2 lg:text-sm { entry.checked ? 'opacity-50' : '' }">
             Am günstigsten bei
             <strong class="text-green-dark font-semibold">{merchant}</strong>für
-            <strong class="text-green-dark font-semibold">{price.price} €</strong>
+            <strong class="text-green-dark font-semibold">{price.price ?? 0} €</strong>
         </p>
     {/if}
 </li>
